dbtest: Refactor dsn construction
